PhaseUtils.jl

Documentation for PhaseUtils.jl

Types and Functions

PhaseUtils.ap2mask — Method

ap2mask(ap)

Converts ap array to mask, so that every zrro is mapped to NaN and non-zero elements to 1.

See also mask2ap.

PhaseUtils.bboxview — Function

bboxview(arr, mask, pad = 0)

Make a box corresponding to the mask not-NaN elements surrounded by pad.

PhaseUtils.circlemask! — Method

circlemask!(a::Matrix{<:Number}, [cx, cy,] r)

Set to zero values of matrix a outside the circle with center (cx,cy) and radius r.

If omitted, (cx,cy) is the center of the matrix.

PhaseUtils.circlemask — Method

circlemask(dims::NTuple{2, Int}, cx, cy, r)

Create a boolean matrix of size dims with true only inside the circle with center (cx,cy) and radius r.

If omitted, (cx,cy) is the center of the domain: (dims+1)/2.

Examples

julia> circlemask((6,8), 2.5,3, 1.5)
6×8 Matrix{Bool}:
 0  0  1  0  0  0  0  0
 0  1  1  1  0  0  0  0
 0  1  1  1  0  0  0  0
 0  0  1  0  0  0  0  0
 0  0  0  0  0  0  0  0
 0  0  0  0  0  0  0  0

julia> circlemask((10,10), 3)
10×10 Matrix{Bool}:
 0  0  0  0  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0
 0  0  0  1  1  1  1  0  0  0
 0  0  1  1  1  1  1  1  0  0
 0  0  1  1  1  1  1  1  0  0
 0  0  1  1  1  1  1  1  0  0
 0  0  1  1  1  1  1  1  0  0
 0  0  0  1  1  1  1  0  0  0
 0  0  0  0  0  0  0  0  0  0
 0  0  0  0  0  0  0  0  0  0

PhaseUtils.crop — Method

crop(a::AbstractArray, cropsize[, cropcenter])
crop(cropsize[, cropcenter]) creates a cropping operator.

Crop the array to cropsize around cropcenter. If cropcenter is omitted, use central element of the array.

cropsize and cropcenter can be either tuples/vectors of integers or just single integer, which represent a tuple of identical numbers.

Examples

julia> a = reshape(1:15, 3, 5); crop(a, 1, (1, 3))
1×1 Matrix{Int64}:
 7

julia> crop(a, (2, 2), (1, 3))
1×2 Matrix{Int64}:
 4  7

julia> crop(3)(a)
3×3 Matrix{Int64}:
 4  7  10
 5  8  11
 6  9  12

PhaseUtils.getresmap — Method

getresmap(gx, gy) -> resmap

Find residues in the phase, which correspond to the residues in the wrapped phase gradients expressed in 2pi multiples.

PhaseUtils.getresmap — Method

getresmap(phase) -> resmap

Find residues in the phase, which correspond to the residues in the wrapped phase gradients expressed in 2pi multiples.

PhaseUtils.getresmapsparce — Method

getresmapsparce(gx,gy) -> posx, posy, resmap

Find residues in the (inconsistent) gradient field.

PhaseUtils.getresmapsparce — Method

getresmapsparce(phase) -> posx, posy, resmap

Find residues in the phase, which correspond to the residues in the wrapped phase gradients expressed in 2pi multiples.

PhaseUtils.integrate_2dgrad — Method

integrate_2dgrad(gx, gy[, gradmethod=default_grad_method(gx)])

TBW

PhaseUtils.integrate_periodic_grad — Method

integrate_periodic_grad(g)

Integrate 1D gradient with periodic boundary conditions

PhaseUtils.itoh — Method

itoh(phi)

Itoh's algorithm for 1D phase unwrapping.

Example

julia> ph = 1:10
1:10

julia> ph_w = phwrap(ph)
10-element Vector{Float64}:
  1.0
  2.0
  3.0
 -2.2831853071795867
 -1.2831853071795865
 -0.28318530717958645
  0.7168146928204135
  1.7168146928204135
  2.7168146928204133
 -2.566370614359173

julia> itoh(ph_w)
10-element Vector{Float64}:
  1.0
  2.0
  3.0
  4.0
  5.0
  6.0
  7.0
  8.0
  9.0
 10.0

PhaseUtils.linearphase — Method

linearphase(T=Float64, dims::NTuple{2,Int},  cx, cy, kx, ky)

Create a matrix with linear phase with slopes (kx, ky) and having zero at point (cx, cy).

Example

julia> linearphase((5,7), 3, 3, 1.5, 1.5)
5×7 Matrix{Float64}:
 -6.0  -4.5  -3.0  -1.5  0.0  1.5  3.0
 -4.5  -3.0  -1.5   0.0  1.5  3.0  4.5
 -3.0  -1.5   0.0   1.5  3.0  4.5  6.0
 -1.5   0.0   1.5   3.0  4.5  6.0  7.5
  0.0   1.5   3.0   4.5  6.0  7.5  9.0

PhaseUtils.mask2ap — Method

mask2ap(mask)

Converts mask array where points outside the aperture are defines as NaN to a Float array, so that NaN -> 0, not NaN -> 1.

PhaseUtils.membrane_sor! — Method

membrane_sor!(u, f::Array{Real}, mask_internal::Array{Bool})

Solves Poisson equation Δ u = f, with boundary conditions set by u(x) for x ∉ mask_internal using Successive Overrelaxation method (see Numerical Recipes, ch.20.5.1).

PhaseUtils.membrane_sor — Method

membrane_sor(f::Array{Real}, mask_internal::Array{Bool})

Solves Poisson equation Δ u = f, with boundary conditions u(x) = 0 <=> x ∉ mask_internal using Successive Overrelaxation method (see Numerical Recipes, ch.20.5.1).

PhaseUtils.phwrap — Method

phwrap(ϕ)

Wrap phase ϕ: ψ = phwrap(ϕ) ⇔ ψ = ϕ + 2πk, k∈Z, -π <ψ ≤π. If called on an array, works pointwise.

PhaseUtils.toArray — Method

toArray(ind<:Array{CartesianIndex}, vals; crop = true, shift=0, defaultel = 0) --> A

Construct an array A for which A[ind[i]] = val[i]. If crop is true, return only the minimal bounding box of A. shifted to shift pixels.

PhaseUtils.unwrap — Function

unwrap(phase, aperture, method = LS(); restore_piston=false)

Unwrap 2D phase defined inside aperture using method. The values outside the aperture are ignored. If the aperture argument is omitted, it is calucalated automatically from the values of phase array: every pixel with not NaN or missing value is considered to be inside the aperture.

Methods

LS(decomposition, integration): uses the least-squares decomposition of the wrapped gradient of the wrapped phase in the rotor-free and solenodial fields and integration of the rotor-free part. By default:

Talmi-Ribak for decompositon

F1F2 for integration

PhaseUtils.unwrap_LS — Method

unwrap_LS(phase, aperture; restore_piston=false)

Unwrap 2D phase defined inside aperture using the Least-Squares decomposition of the wrapped gradient of the wprapped phase in the rotor-free and solenodial field and integration of the rotor-free part.

Index

PhaseUtils.ap2mask

PhaseUtils.bboxview

PhaseUtils.circlemask

PhaseUtils.circlemask!

PhaseUtils.crop

PhaseUtils.getresmap

PhaseUtils.getresmap

PhaseUtils.getresmapsparce

PhaseUtils.getresmapsparce

PhaseUtils.integrate_2dgrad

PhaseUtils.integrate_periodic_grad

PhaseUtils.itoh

PhaseUtils.linearphase

PhaseUtils.mask2ap

PhaseUtils.membrane_sor

PhaseUtils.membrane_sor!

PhaseUtils.phwrap

PhaseUtils.toArray

PhaseUtils.unwrap

PhaseUtils.unwrap_LS